+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # v2.5.0 (unreleased)
 ### Notes
-- This release supports OneView Rest API versions 800/1000/1200 for below resources.
+- This release supports OneView Rest API versions 600/800/1000/1200 for below resources.
 
 ### Oneview Features supported
+- Server Certificate
 - Hypervisor Manager
 
 ### Bug fixes & Enhancements: